builder-agent/pyproject.toml
[project]
name = "skill-builder-agent"
version = "0.2.0"
description = "ADK multi-agent pipeline that generates SKILL.md files from natural language, with Neo4j vector-powered exemplar retrieval."
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0,<2.0.0",
    "litellm>=1.40.0,<3.0.0",
    "uvicorn>=0.30.0",
    "starlette>=0.40.0,<1.0.0",
    "pydantic-settings>=2.8.1",
    "neo4j>=5.20.0,<6.0.0",
    "sentence-transformers>=3.0.0",
    "sse-starlette>=2.0.0",
    "pyyaml>=6.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation-starlette>=0.41b0",
    "pydantic>=2.0,<3.0",
    "networkx>=3.0",
    "python-louvain>=0.16",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "httpx>=0.27",
]

[project.scripts]
skill-builder-server = "skill_builder.server:run"
skill-builder = "skill_builder.cli:main"

[project.urls]
Homepage = "https://github.com/rrbanda/skill-mp"
Repository = "https://github.com/rrbanda/skill-mp"
Issues = "https://github.com/rrbanda/skill-mp/issues"
Documentation = "https://github.com/rrbanda/skill-mp#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/skill_builder"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--cov=skill_builder --cov-report=term-missing --cov-fail-under=50"

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
